feat: add file upload with user ownership
- Add user_id to files table with migration
- Update Files schema with user association
- Update Storage context to filter by user
- Update FileController to use current_user for all operations
- Update frontend api.js with credentials for session cookies
- Update uploader to use Phoenix API

+++ personal_drive_api/lib/personal_drive_api_web/controllers/file_controller.ex
@@ -7,26 +7,44 @@ defmodule PersonalDriveApiWeb.FileController do
 
   action_fallback PersonalDriveApiWeb.FallbackController
 
+  # Get current user from Pow assigns
+  defp current_user(conn) do
+    conn.assigns[:current_user]
+  end
+
   # List files (root or folder)
   def index(conn, %{"parent_id" => parent_id}) do
-    files = Storage.list_files_by_parent(parent_id)
+    user = current_user(conn)
+    files = Storage.list_files_by_parent(parent_id, user.id)
     render(conn, :index, files: files)
   end
 
   def index(conn, _params) do
-    files = Storage.list_root_files()
+    user = current_user(conn)
+    files = Storage.list_root_files(user.id)
     render(conn, :index, files: files)
   end
 
   # Get single file/folder
   def show(conn, %{"id" => id}) do
+    user = current_user(conn)
     file = Storage.get_file!(id)
-    render(conn, :show, file: file)
+    
+    # Verify ownership
+    if file.user_id == user.id do
+      render(conn, :show, file: file)
+    else
+      conn
+      |> put_status(:forbidden)
+      |> json(%{error: "Not authorized"})
+    end
   end
 
   # Create folder
   def create_folder(conn, %{"name" => name, "parent_id" => parent_id}) do
-    case Storage.create_folder(name, parent_id) do
+    user = current_user(conn)
+    
+    case Storage.create_folder(name, user.id, parent_id) do
       {:ok, folder} ->
         conn
         |> put_status(:created)
@@ -45,26 +63,24 @@ defmodule PersonalDriveApiWeb.FileController do
 
   # Upload file
   def upload(conn, %{"file" => file, "parent_id" => parent_id}) do
-    upload_file(conn, file, parent_id)
+    user = current_user(conn)
+    upload_file(conn, file, parent_id, user.id)
   end
 
   def upload(conn, %{"file" => file}) do
-    upload_file(conn, file, nil)
+    user = current_user(conn)
+    upload_file(conn, file, nil, user.id)
   end
 
-  defp upload_file(conn, file, parent_id) do
+  defp upload_file(conn, file, parent_id, user_id) do
     filename = file.filename
     content_type = file.content_type
-    # Read from the uploaded file path
     {:ok, binary} = File.read(file.path)
 
-    # Generate R2 key
     key = R2.generate_key(parent_id, filename)
 
-    # Upload to R2
     case R2.upload_binary(binary, key, content_type) do
       {:ok, r2_data} ->
-        # Save to database
         case Storage.create_file(%{
               name: filename,
               size: byte_size(binary),
@@ -72,7 +88,8 @@ defmodule PersonalDriveApiWeb.FileController do
               r2_key: r2_data.r2_key,
               r2_etag: r2_data.r2_etag,
               is_folder: false,
-              parent_id: parent_id
+              parent_id: parent_id,
+              user_id: user_id
             }) do
           {:ok, file_record} ->
             conn
@@ -80,7 +97,6 @@ defmodule PersonalDriveApiWeb.FileController do
             |> render(:show, file: file_record)
 
           {:error, changeset} ->
-            # Rollback R2 upload on DB error
             R2.delete_file(key)
             conn
             |> put_status(:unprocessable_entity)
@@ -94,7 +110,7 @@ defmodule PersonalDriveApiWeb.FileController do
     end
   end
 
-  # Get signed upload URL (for direct browser uploads)
+  # Get signed upload URL
   def presigned_upload(conn, %{"filename" => filename, "content_type" => content_type, "parent_id" => parent_id}) do
     key = R2.generate_key(parent_id, filename)
 
@@ -109,8 +125,10 @@ defmodule PersonalDriveApiWeb.FileController do
     end
   end
 
-  # Confirm upload after client uploads to R2
+  # Confirm upload
   def confirm_upload(conn, %{"key" => key, "name" => name, "size" => size, "content_type" => content_type, "parent_id" => parent_id}) do
+    user = current_user(conn)
+    
     case Storage.create_file(%{
           name: name,
           size: size,
@@ -118,7 +136,8 @@ defmodule PersonalDriveApiWeb.FileController do
           r2_key: key,
           r2_etag: "pending",
           is_folder: false,
-          parent_id: parent_id
+          parent_id: parent_id,
+          user_id: user.id
         }) do
       {:ok, file_record} ->
         conn
@@ -134,47 +153,59 @@ defmodule PersonalDriveApiWeb.FileController do
 
   # Get signed download URL
   def download_url(conn, %{"id" => id}) do
+    user = current_user(conn)
     file = Storage.get_file!(id)
 
-    if file.is_folder do
+    if file.user_id != user.id do
       conn
-      |> put_status(:bad_request)
-      |> json(%{error: "Cannot download folder"})
+      |> put_status(:forbidden)
+      |> json(%{error: "Not authorized"})
     else
-      case R2.get_signed_url(file.r2_key) do
-        {:ok, url} ->
-          json(conn, %{download_url: url})
+      if file.is_folder do
+        conn
+        |> put_status(:bad_request)
+        |> json(%{error: "Cannot download folder"})
+      else
+        case R2.get_signed_url(file.r2_key) do
+          {:ok, url} ->
+            json(conn, %{download_url: url})
 
-        {:error, error} ->
-          conn
-          |> put_status(:unprocessable_entity)
-          |> json(%{error: "Failed to get download URL: #{inspect(error)}"})
+          {:error, error} ->
+            conn
+            |> put_status(:unprocessable_entity)
+            |> json(%{error: "Failed to get download URL: #{inspect(error)}"})
+        end
       end
     end
   end
 
   # Delete file/folder
   def delete(conn, %{"id" => id}) do
+    user = current_user(conn)
     file = Storage.get_file!(id)
 
-    # If it's a file, delete from R2
-    unless file.is_folder do
-      R2.delete_file(file.r2_key)
-    end
+    if file.user_id != user.id do
+      conn
+      |> put_status(:forbidden)
+      |> json(%{error: "Not authorized"})
+    else
+      unless file.is_folder do
+        R2.delete_file(file.r2_key)
+      end
 
-    # If it's a folder, delete all children recursively
-    if file.is_folder do
-      Storage.delete_folder_recursive(id)
-    end
+      if file.is_folder do
+        Storage.delete_folder_recursive(id)
+      end
 
-    case Storage.delete_file(file) do
-      {:ok, _} ->
-        send_resp(conn, :no_content, "")
+      case Storage.delete_file(file) do
+        {:ok, _} ->
+          send_resp(conn, :no_content, "")
 
-      {:error, changeset} ->
-        conn
-        |> put_status(:unprocessable_entity)
-        |> render(:error, changeset: changeset)
+        {:error, changeset} ->
+          conn
+          |> put_status(:unprocessable_entity)
+          |> render(:error, changeset: changeset)
+      end
     end
   end
 end
